+++ src/main/java/com/squad03/flap/DTO/DetalheTarefa.java
@@ -13,7 +13,7 @@ public record DetalheTarefa(
         String descricao,
         StatusTarefa status,
         PrioridadeTarefa prioridade,
-        Integer posicao,
+        Double posicao,
         LocalDateTime dtCriacao,
         LocalDateTime dtEntrega,
         LocalDateTime dtConclusao,
